removed all theme js and html to theme files as new page
