+++ packages/ui/src/index.js
@@ -70,7 +70,6 @@ Object.assign(d3, d3cam02, d3hsluv, d3hsv, d33d);
 
 import * as charts from './charts.js';
 import * as chartData from './data.js';
-import * as theme from './theme.js';
 
 var bgFieldInput = document.getElementById('bgField');
 var background = bgFieldInput.value;
@@ -132,6 +131,7 @@ function throttle(func, wait) {
     }
   };
 }
+exports.throttle = throttle;
 
 function paramSetup() {
   colorspaceOptions();
@@ -373,6 +373,7 @@ function deleteColor(e) {
   self.remove();
   colorInput();
 }
+exports.deleteColor = deleteColor;
 
 window.openTab = function openTab(evt, tabName) {
   // Declare all variables
@@ -426,6 +427,7 @@ document.getElementById("tabTheme").addEventListener('click', colorInput);
 function randomId() {
    return Math.random().toString(36).replace(/[^a-z]+/g, '').substr(2, 10);
 }
+exports.randomId = randomId;
 
 function createDemo(c, z) {
   var smallText = 'Small text demo';
@@ -934,385 +936,3 @@ function updateCharts(selectObject) {
   colorSpaceFeedback(spaceOpt);
 }
 window.updateCharts = updateCharts;
-
-/* ---------------------------  */
-/*     CREATE THEME SCRIPTS     */
-/* ---------------------------
-1. Generate Color Scale object
-2. For each color scale object, function for color inputs.
-3. AddColor function (per object)
-4. Delete object
-5. Add object (run generate color)
-6. Param setup
-7. Param input (read if has params )
-8. Function output
-9. Colors output (for each object, output colors in panel)
-10. Toggle configs visibility
-11. Brightness slider value sync
-12. Base color input options based on color objects available
-13. Add color from Leonardo URL
-14.
-*/
-
-window.addColorScale = addColorScale;
-function addColorScale() {
-  // create unique ID for color object
-  let thisId = randomId();
-  // generate color input objects:
-  // gradient, inputs, etc.
-  let wrapper = document.getElementById('themeColorWrapper');
-  // Remove empty state
-  if(!document.getElementById('themeColorEmptyState')) {
-    // Do nothing
-  } else {
-    wrapper.innerHTML = ' ';
-  }
-
-  // Create theme item
-  let item = document.createElement('div');
-  item.className = 'themeColor_item';
-  item.id = thisId;
-
-  // Create gradient
-  let gradient = document.createElement('div');
-  gradient.className = 'themeColor_gradient';
-  let gradientId = thisId.concat('_gradient');
-  gradient.id = gradientId;
-
-  // Create form container
-  let inputs = document.createElement('div');
-  inputs.className = 'spectrum-Form spectrum-Form--row themeColor_configs';
-
-  // Color Name Input
-  let colorName = document.createElement('div');
-  colorName.className = 'spectrum-Form-item';
-  let colorNameLabel = document.createElement('label');
-  colorNameLabel.className = 'spectrum-FieldLabel';
-  colorNameLabel.for = thisId.concat('_colorName');
-  let colorNameInput = document.createElement('input');
-  colorNameInput.type = 'text';
-  colorNameInput.className = 'spectrum-Textfield colorNameInput';
-  colorNameInput.id = thisId.concat('_colorName');
-  colorNameInput.name = thisId.concat('_colorName');
-  colorNameInput.placeholder = 'My Color';
-  colorNameInput.onchange = throttle(themeInput, 10);
-  colorNameLabel.innerHTML = 'Color name';
-  colorName.appendChild(colorNameLabel);
-  colorName.appendChild(colorNameInput);
-
-  // Key Color Input
-  let keyColors = document.createElement('div');
-  keyColors.className = 'spectrum-Form-item';
-  let keyColorsLabel = document.createElement('label');
-  keyColorsLabel.className = 'spectrum-FieldLabel';
-  keyColorsLabel.for = thisId.concat('_keyColors');
-  let keyColorsInput = document.createElement('div');
-  keyColorsInput.className = 'keyColorsWrapper';
-  let keyColorsId = thisId.concat('_keyColors');
-  keyColorsInput.id = keyColorsId;
-  keyColorsLabel.innerHTML = 'Key colors';
-  keyColors.appendChild(keyColorsLabel);
-  keyColors.appendChild(keyColorsInput);
-
-  // Key Colors Actions
-  let addColors = document.createElement('div');
-  addColors.className = 'spectrum-Form-item';
-  let addButton = document.createElement('button');
-  addButton.className = 'spectrum-ActionButton';
-  let buttonId = thisId.concat('_addKeyColor');
-  addButton.id = buttonId;
-  addButton.onclick = themeAddColor;
-  addButton.innerHTML = `
-  <svg xmlns:xlink="http://www.w3.org/1999/xlink" class="spectrum-Icon spectrum-Icon--sizeS" focusable="false" aria-hidden="true" aria-label="Add">
-    <use xlink:href="#spectrum-icon-18-Add" />
-  </svg>
-  `;
-  addColors.appendChild(addButton);
-
-  // Interpolation mode
-  let interp = document.createElement('div');
-  interp.className = 'spectrum-Form-item';
-  let interpLabel = document.createElement('label');
-  interpLabel.className = 'spectrum-FieldLabel';
-  interpLabel.for = thisId.concat('_mode');
-  let interpLabelText = 'Colorspace interpolation';
-  let interpDropdown = document.createElement('div');
-  interpDropdown.className = 'spectrum-Dropdown';
-  interpDropdown.id = thisId.concat('_modeDropdown');
-  let interpSelect = document.createElement('select');
-  interpSelect.className = 'spectrum-FieldButton spectrum-Dropdown-trigger';
-  interpSelect.id = thisId.concat('_mode');
-  interpSelect.name = thisId.concat('_mode');
-  interpSelect.oninput = throttle(themeInput, 20);
-  let interpDropdownIcon = document.createElement('span');
-  interpDropdownIcon.className = 'spectrum-Dropdown-iconWrapper';
-  interpDropdownIcon.innerHTML = `
-  <svg xmlns:xlink="http://www.w3.org/1999/xlink" class="spectrum-Icon spectrum-UIIcon-ChevronDownMedium spectrum-Dropdown-icon">
-    <use xlink:href="#spectrum-css-icon-ChevronDownMedium"></use>
-  </svg>`;
-
-  interpLabel.innerHTML = interpLabelText;
-  interpDropdown.appendChild(interpSelect);
-  interpDropdown.appendChild(interpDropdownIcon);
-  interp.appendChild(interpLabel);
-  interp.appendChild(interpDropdown);
-
-  // Interpolation options
-  interpSelect.options.length = 0;
-
-  let opts = {
-    'CAM02': 'CIECAM02',
-    'LCH': 'Lch',
-    'LAB': 'Lab',
-    'HSL': 'HSL',
-    'HSLuv': 'HSLuv',
-    'HSV': 'HSV',
-    'RGB': 'RGB'
-  };
-
-  for(let index in opts) { interpSelect.options[interpSelect.options.length] = new Option(opts[index], index); }
-  interpSelect.value = 'CAM02';
-
-  // Ratios
-  let ratios = document.createElement('div');
-  ratios.className = 'spectrum-Form-item';
-  let ratiosLabel = document.createElement('label');
-  ratiosLabel.className = 'spectrum-FieldLabel';
-  ratiosLabel.innerHTML = 'Contrast ratios';
-  ratiosLabel.for = thisId.concat('_ratios');
-  let ratiosInput = document.createElement('input');
-  ratiosInput.type = 'text';
-  ratiosInput.placeholder = '3, 4.5';
-  ratiosInput.className = 'spectrum-Textfield';
-  ratiosInput.id = thisId.concat('_ratios');
-  ratiosInput.name = thisId.concat('_ratios');
-  ratiosInput.oninput = throttle(themeInput, 10);
-  ratios.appendChild(ratiosLabel);
-  ratios.appendChild(ratiosInput);
-
-  // Actions
-  let actions = document.createElement('div');
-  actions.className = 'spectrum-ButtonGroup spectrum-Form-item';
-  let edit = document.createElement('button');
-  edit.className = 'spectrum-ActionButton';
-  edit.innerHTML = `
-  <!-- <span class="spectrum-ActionButton-label">Add from URL</span> -->
-  <svg xmlns:xlink="http://www.w3.org/1999/xlink" class="spectrum-Icon spectrum-Icon--sizeS" focusable="false" aria-hidden="true" aria-label="Add">
-    <use xlink:href="#spectrum-icon-18-Edit" />
-  </svg>`
-  // edit.addEventListener('click', openEditColorScale) // TODO => create openEditColorScale function to open colors tab w/ settings of this object.
-  let deleteColor = document.createElement('button');
-  deleteColor.className = 'spectrum-ActionButton';
-  deleteColor.id = thisId.concat('_delete');
-  deleteColor.innerHTML = `
-  <!-- <span class="spectrum-ActionButton-label">Add Color</span> -->
-  <svg xmlns:xlink="http://www.w3.org/1999/xlink" class="spectrum-Icon spectrum-Icon--sizeS" focusable="false" aria-hidden="true" aria-label="Add">
-    <use xlink:href="#spectrum-icon-18-Delete" />
-  </svg>`;
-  deleteColor.onclick = themeDeleteItem;
-  actions.appendChild(edit);
-  actions.appendChild(deleteColor);
-
-  // Put it all together
-  inputs.appendChild(colorName);
-  inputs.appendChild(keyColors);
-  inputs.appendChild(addColors)
-  inputs.appendChild(interp);
-  inputs.appendChild(ratios);
-  inputs.appendChild(actions);
-
-  item.appendChild(gradient);
-  item.appendChild(inputs);
-
-  wrapper.appendChild(item);
-
-  // Then run functions on the basic placeholder inputs
-  document.getElementById(buttonId).click();
-  // generate the number of values equal to the width of the item
-  let n = window.innerWidth - 272;
-  let rampData = contrastColors.createScale({swatches: n, colorKeys: ['#000000'], colorspace: 'LAB'});
-  let colors = rampData.colors;
-
-  themeRamp(colors, n, gradientId);
-  document.getElementById(thisId.concat('_colorName')).addEventListener('input', baseScaleOptions);
-  document.getElementById(thisId.concat('_delete')).addEventListener('click', themeDeleteItem);
-}
-
-function themeRamp(colors, n = window.innerWidth - 272, dest) {
-  let container = document.getElementById(dest);
-  container = d3.select(container);
-
-  let canvas = container.append("canvas")
-    .attr("width", n)
-    .attr("height", 1);
-  let context = canvas.node().getContext("2d");
-
-  canvas.style.height = "32px";
-  canvas.style.width = n;
-  canvas.style.imageRendering = "pixelated";
-  canvas.className = 'themeColor_canvas';
-  for (let i = 1; i < n; ++i) {
-    context.fillStyle = colors[i];
-    context.fillRect(i, 0, 1, 32);
-  }
-  return canvas;
-}
-
-
-// Recreation of addColor function, specifying items needed for theme UI
-function themeAddColor(c) {
-  let thisId = this.id;
-  let parent = thisId.replace('_addKeyColor', '');
-  let destId = parent.concat('_keyColors');
-  let dest = document.getElementById(destId);
-  let div = document.createElement('div');
-
-  let randId = randomId();
-  div.className = 'keyColor';
-  div.id = randId + '-item';
-  let sw = document.createElement('input');
-  sw.type = "color";
-  sw.value = c;
-  sw.oninput = throttle(themeInput, 50);
-
-  sw.className = 'keyColor-Item';
-  sw.id = randId + '-sw';
-  sw.style.backgroundColor = c;
-
-  let button = document.createElement('button');
-  button.className = 'spectrum-ActionButton';
-  button.innerHTML = `
-  <svg class="spectrum-Icon spectrum-Icon--sizeS" focusable="false" aria-hidden="true" aria-label="Delete">
-    <use xlink:href="#spectrum-icon-18-Delete" />
-  </svg>`;
-
-  button.onclick = deleteColor;
-  div.appendChild(sw);
-  div.appendChild(button);
-  dest.appendChild(div);
-}
-
-// TODO: Get this to work.
-function themeDeleteItem(e) {
-  var id = e.target.parentNode.id;
-  var self = document.getElementById(id);
-
-  self.remove();
-  baseScaleOptions();
-  themeInput();
-}
-
-// Create options for colors to use as base scale
-function baseScaleOptions() {
-  let baseSelect = document.getElementById('themeBase');
-  baseSelect.options.length = 0;
-
-  let colorNames = document.getElementsByClassName('colorNameInput');
-  let opts = {};
-  for (let i = 0; i < colorNames.length; i++) {
-    let colorname = colorNames[i].value;
-    opts[colorname] = colorname;
-  }
-
-  for(let index in opts) { baseSelect.options[baseSelect.options.length] = new Option(opts[index], index); }
-}
-
-var themeConfigs = {};
-
-window.themeInput = themeInput;
-function themeInput() {
-  // Run this function any time an object is added or any item is changed.
-  // Gather all color scale objects
-  let items = document.getElementsByClassName('themeColor_item');
-
-  let baseScale = {};
-  let colorScales = [];
-
-  // Find name of selected base scale color
-  let baseSelect = document.getElementById('themeBase');
-  let baseSelectValue = baseSelect.value;
-  // Find which object has this name
-  let colorNames = document.getElementsByClassName('colorNameInput');
-  // let str = 'input[value="'.concat(baseSelectValue).concat('"]');
-  // let test = document.querySelectorAll( str );
-  // let colorNameId = test.id;
-  // console.log(colorNameId);
-
-  // let baseScale = {
-  //   colorKeys: ['#cacaca'],
-  //   colorspace: 'HSL'
-  // },
-
-  for (let i = 0; i < items.length; i++) {
-    let id = items[i].id;
-    let thisElement = document.getElementById(id);
-    // 1. find color name
-    let colorNameInput = id.concat('_colorName');
-    let colorName = document.getElementById(colorNameInput).value;
-    // 2. find all key colors
-    let colorKeys = thisElement.getElementsByClassName('keyColor-Item');
-    let inputColors = [];
-    let tempArgs = [];
-    for(let i=0; i<colorKeys.length; i++) {
-      inputColors.push(colorKeys[i].value);
-    }
-    // Convert input value into a split array of hex values.
-    // remove any whitespace from inputColors
-    tempArgs.push(inputColors);
-    let colorArgs = tempArgs.join("").split(',').filter(String);
-
-    // 3. find mode
-    let modeId = id.concat('_mode');
-    let modeSelect = document.getElementById(modeId);
-    let mode = modeSelect.value;
-
-    // 4. find ratios
-    let ratiosId = id.concat('_ratios');
-    let ratiosInput = document.getElementById(ratiosId);
-    let rVals = ratiosInput.value;
-    let r = new Array(rVals);
-    let rSplit = r.join("").split(',');
-    let ratios = rSplit.map(x => parseFloat(x));
-    // TODO: remove all values of NaN
-
-    // Re generate gradient
-    let gradientId = id.concat('_gradient');
-    let gradient = document.getElementById(gradientId);
-    gradient.innerHTML = ' ';
-    let n = window.innerWidth - 272;
-    let rampData = contrastColors.createScale({swatches: n, colorKeys: colorArgs, colorspace: mode});
-    let colors = rampData.colors;
-    themeRamp(colors, n, gradientId);
-    // - assign properties to an object
-
-    let colorObj = {
-      name: colorName,
-      colorKeys: colorArgs,
-      colorspace: mode,
-      ratios: ratios
-    };
-
-    colorScales.push(colorObj);
-  }
-  // Then, pass all props to 'generateAdaptiveTheme'
-  // For each returned object
-  // display swatch on right panel ?
-  themeConfigs.baseScale = baseScale;
-  themeConfigs.colorScales = colorScales;
-
-  console.log(themeConfigs);
-}
-
-window.toggleConfigs = toggleConfigs;
-function toggleConfigs(e) {
-  let configs = document.getElementsByClassName('themeColor_configs');
-  let button = document.getElementById('toggleConfigButton');
-  let gradient = document.getElementsByClassName('themeColor_gradient');
-  button.classList.toggle('is-selected');
-
-  for (let i = 0; i < configs.length; i ++) {
-    configs[i].classList.toggle('is-hidden');
-    gradient[i].classList.toggle('is-large');
-  }
-
-}
